refactor: move share action to preview and cleanup
- Move share functionality from  to  as
- Remove unused  component and  function
- Cleanup unused imports in
- Update  to include  in all preview variants
