+++ src/components/prompt-wizard/NavigationActions.tsx
@@ -1,38 +1,20 @@
-import { useState } from "react";
 import { Link } from "@tanstack/react-router";
-import { Plus, Sparkles, Copy, Share2, Loader2, Check } from "lucide-react";
-import { useMutation } from "convex/react";
-import { toast } from "sonner";
+import { Plus, Sparkles } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
 
 import { useTrackMixpanel } from "@/utils/analytics/MixpanelProvider";
 import { useWizardStore } from "@/stores/wizard-store";
-import { getOrCreateSessionId } from "@/utils/session";
-import { PromptWizardData } from "@/utils/prompt-wizard/schema";
-
-// @ts-ignore - Valid after npx convex dev
-import { api } from "../../../convex/_generated/api";
 
 interface NavigationActionsProps {
   page: string;
-  wizardData?: PromptWizardData;
 }
 
 /**
- * Navigation actions: Back To Home + Create New Prompt + Share
+ * Navigation actions: Back To Home + Create New Prompt
  * Reusable across wizard and share pages
  */
-export function NavigationActions({ page, wizardData }: NavigationActionsProps) {
+export function NavigationActions({ page }: NavigationActionsProps) {
   const trackEvent = useTrackMixpanel();
   const reset = useWizardStore((state) => state.reset);
 
@@ -55,8 +37,6 @@ export function NavigationActions({ page, wizardData }: NavigationActionsProps) 
 
       {/* Actions */}
       <div className="flex items-center gap-3">
-        {wizardData && <ShareButton wizardData={wizardData} page={page} />}
-
         <Button
           size="sm"
           asChild
@@ -73,103 +53,5 @@ export function NavigationActions({ page, wizardData }: NavigationActionsProps) 
         </Button>
       </div>
     </div>
-  );
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// SHARE BUTTON COMPONENT
-// ─────────────────────────────────────────────────────────────────────────────
-
-function ShareButton({ wizardData, page }: { wizardData: PromptWizardData; page: string }) {
-  const savePrompt = useMutation(api.prompts.savePrompt);
-  const trackEvent = useTrackMixpanel();
-
-  const [isOpen, setIsOpen] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const [shareLink, setShareLink] = useState("");
-  const [copied, setCopied] = useState(false);
-
-  // Consider a prompt active if it has at least user interaction
-  const hasContent = wizardData.updatedAt > 0;
-
-  const handleCreateLink = async () => {
-    if (!hasContent) {
-      toast.error("Add some content to your prompt first!");
-      return;
-    }
-    setIsLoading(true);
-    try {
-      const sessionId = getOrCreateSessionId();
-      const { slug } = await savePrompt({ promptData: wizardData, sessionId });
-
-      const url = `${window.location.origin}/prompts/${slug}`;
-      setShareLink(url);
-
-      trackEvent("cta_clicked_get_shareable_link", {
-        page,
-        slug,
-        role: wizardData.ai_role,
-      });
-    } catch (error) {
-      console.error(error);
-      toast.error("Failed to create share link");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleCopy = () => {
-    navigator.clipboard.writeText(shareLink);
-    setCopied(true);
-    setTimeout(() => setCopied(false), 2000);
-    toast.success("Link copied!");
-  };
-
-  if (!hasContent) return null;
-
-  return (
-    <Dialog open={isOpen} onOpenChange={setIsOpen}>
-      <DialogTrigger asChild>
-        <Button
-          size="sm"
-          variant="secondary"
-          className="uppercase font-bold text-xs shadow-[2px_2px_0px_0px_hsl(var(--foreground))] border-2 border-foreground hover:translate-x-[2px] hover:translate-y-[2px] hover:shadow-none transition-all"
-          onClick={handleCreateLink}
-        >
-          <Share2 className="w-4 h-4 mr-2" />
-          Share Link
-        </Button>
-      </DialogTrigger>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Get Shareable Link</DialogTitle>
-          <DialogDescription>
-            Create a permanent, SEO-optimized URL for this prompt to share on social media.
-          </DialogDescription>
-        </DialogHeader>
-
-        <div className="flex flex-col gap-4 py-4">
-          {isLoading ? (
-            <div className="flex flex-col items-center justify-center py-8 gap-2 text-muted-foreground">
-              <Loader2 className="w-8 h-8 animate-spin" />
-              <p className="text-sm">Creating your unique link...</p>
-            </div>
-          ) : (
-            <div className="flex items-center gap-2">
-              <Input
-                value={shareLink || "Click Share to generate link"}
-                readOnly
-                className="font-mono text-xs"
-              />
-              {shareLink && (
-                <Button size="icon" onClick={handleCopy}>
-                  {copied ? <Check className="w-4 h-4" /> : <Copy className="w-4 h-4" />}
-                </Button>
-              )}
-            </div>
-          )}
-        </div>
-      </DialogContent>
-    </Dialog>
   );
 }
