+++ src/routes/index.tsx
@@ -2,7 +2,6 @@ import { createFileRoute } from "@tanstack/react-router";
 
 import { ChatLandingPage } from "@/components/landing/v2/ChatLandingPage"; // v2
 import { ErrorComponentWithSentry } from "@/components/ErrorComponentWithSentry";
-import { TestGeminiButton } from "@/components/TestGeminiButton";
 
 // ═══════════════════════════════════════════════════════════════════════════
 // ROUTE COMPONENT
@@ -19,14 +18,6 @@ import { TestGeminiButton } from "@/components/TestGeminiButton";
  * 5. Default - v1
  */
 function LandingRouter() {
-  // return (
-  //   <div className="relative">
-  //     <div className="fixed bottom-4 right-4 z-50">
-  //       <TestGeminiButton />
-  //     </div>
-  //     <ChatLandingPage />
-  //   </div>
-  // );
   return <ChatLandingPage />;
 }
